\name{divo System Requirements}
\alias{divo System Requirements}

\title{System Requirements}
\description{divo package requires R >=3.1.0; Python >= 2.76 and NumPy.

\strong{Installation of Python and NumPy.} 

Python is already installed in computers with an OSX or a Linux OS.
To install Python in Windows visit \url{https://www.python.org/downloads/}; download the latest version of python and follow the installation instruction. For more information visit \url{https://www.python.org/doc/}

To install NumPy go to \url{http://sourceforge.net/projects/numpy/files/NumPy}, download the latest version of software appropriate for your operating system and follow installation manual. For more information visit \url{http://www.numpy.org}.

In Linux both Python and Numpy may be installed using system package manger e.g. "synaptic" in Debian distribution.

\strong{Setting path on Windows. } 

1. Open System Properties and click Advanced system settings > Advanced tab;\cr
2. Click on Environment Variables, under System Variables, find PATH, and click on it;\cr
3. In the Edit window, modify PATH by adding the location of Python folder. \cr

If you do not have the item PATH, you may select to add a new variable and add PATH as the name and the location of the Python folder.
}
